Create Recommendation model file. Move recommendations structs from UserController file to new Recommendation model file.
